# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/users": {
      "get": {
        "operationId": "listUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "email",
                "createdAt",
                "deletedAt"
              ]
            }
          },
          {
            "name": "emailVerified",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "active"
              ]
            }
          },
          {
            "name": "deleted",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsersResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      },
      "post": {
        "operationId": "createUser",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Chave única gerada pelo cliente por tentativa (ex.: UUID v4). Reenviar a mesma chave com o mesmo payload devolve a resposta original (replay) por até 1h, sem repetir o efeito — seguro pra retry após timeout. Mesma chave com payload diferente retorna 422. Sem o header, o request executa normalmente, sem dedup.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/users/{id}": {
      "put": {
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Chave única gerada pelo cliente por tentativa (ex.: UUID v4). Reenviar a mesma chave com o mesmo payload devolve a resposta original (replay) por até 1h, sem repetir o efeito — seguro pra retry após timeout. Mesma chave com payload diferente retorna 422. Sem o header, o request executa normalmente, sem dedup.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "API da plataforma.\n\n## Autenticação\nSessão via cookie httpOnly (`__Host-rit_session`), emitido por `POST /v1/auth/login`. Enviar cookies em toda chamada autenticada (`credentials: include` no browser). Rotas públicas estão marcadas com `security: []` na operação.\n\n## CSRF — exigido em toda mutação\nRequests não-safe (POST/PATCH/PUT/DELETE) exigem header `Origin` (ou `Referer`) igual à origin do front oficial (`WEB_ORIGIN`); ausente ou divergente → 403. Browsers enviam `Origin` sozinhos; cliente não-browser precisa setar manualmente. Quando os cookies operam com `SameSite=None`, mutações autenticadas exigem também o header `X-CSRF-Token` com o valor do cookie `rit_csrf` (legível por JS, emitido no login).\n\n## Idempotência\nEndpoints marcados aceitam o header opcional `Idempotency-Key` (detalhe na própria operação): mesma chave + mesmo payload → replay da resposta original, sem repetir o efeito.\n\n## Rate limit\nEndpoints com limite devolvem 429 (RFC 7807) com header `Retry-After` em segundos.\n\n## Erros\nRFC 7807 (`application/problem+json`), sempre com `correlationId` pra rastreio.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-rit_session",
        "description": "Cookie de sessão httpOnly emitido pelo login. Nome configurável via COOKIE_NAME (default __Host-rit_session)."
      }
    },
    "schemas": {
...
      "ListUsersResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "emailVerified": {
                  "type": "boolean"
                },
                "accessProfile": {
                  "type": "string",
                  "enum": [
                    "master",
                    "admin",
                    "professional"
                  ]
                },
                "servesClients": {
                  "type": "boolean"
                },
                "permissions": {
                  "maxItems": 24,
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "admin.users.read",
                      "admin.users.create",
                      "admin.users.update",
                      "admin.users.delete",
                      "admin.users.trash.read",
                      "admin.users.trash.restore",
                      "admin.users.trash.purge",
                      "admin.users.access_link.resend",
                      "admin.users.audit.read",
                      "admin.permission_templates.read",
                      "admin.permission_templates.create",
                      "admin.permission_templates.update",
                      "admin.permission_templates.delete",
                      "admin.permission_templates.audit.read",
                      "admin.tags.read",
                      "admin.tags.create",
                      "admin.tags.update",
                      "admin.tags.delete",
                      "admin.tags.trash.read",
                      "admin.tags.trash.restore",
                      "admin.tags.trash.purge",
                      "admin.tags.audit.read",
                      "admin.audit.read",
                      "admin.usage.read"
                    ]
                  }
                },
                "areaIds": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "serviceIds": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "schedulingAreaIds": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "avatarAttachmentId": {
                  "type": "string",
                  "nullable": true
                },
                "createdAt": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "pending",
                    "active"
                  ]
                },
                "accessLinkExpiresAt": {
                  "type": "string",
                  "nullable": true
                },
                "accessLinkExpired": {
                  "type": "boolean"
                },
                "deletedAt": {
                  "type": "string",
                  "nullable": true
                }
              },
              "required": [
                "id",
                "name",
                "email",
                "emailVerified",
                "accessProfile",
                "servesClients",
                "permissions",
                "areaIds",
                "serviceIds",
                "schedulingAreaIds",
                "avatarAttachmentId",
                "createdAt",
                "status",
                "accessLinkExpiresAt",
                "accessLinkExpired",
                "deletedAt"
              ]
            }
          },
          "page": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "page": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "pageSize": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "totalPages": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "total",
              "page",
              "pageSize",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "page"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string"
          },
          "accessProfile": {
            "type": "string",
            "enum": [
              "admin",
              "professional"
            ]
          },
          "servesClients": {
            "default": false,
            "type": "boolean"
          },
          "permissions": {
            "maxItems": 24,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "admin.users.read",
                "admin.users.create",
                "admin.users.update",
                "admin.users.delete",
                "admin.users.trash.read",
                "admin.users.trash.restore",
                "admin.users.trash.purge",
                "admin.users.access_link.resend",
                "admin.users.audit.read",
                "admin.permission_templates.read",
                "admin.permission_templates.create",
                "admin.permission_templates.update",
                "admin.permission_templates.delete",
                "admin.permission_templates.audit.read",
                "admin.tags.read",
                "admin.tags.create",
                "admin.tags.update",
                "admin.tags.delete",
                "admin.tags.trash.read",
                "admin.tags.trash.restore",
                "admin.tags.trash.purge",
                "admin.tags.audit.read",
                "admin.audit.read",
                "admin.usage.read"
              ]
            }
          },
          "areaIds": {
            "default": [],
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          },
          "serviceIds": {
            "default": [],
            "maxItems": 2000,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          },
          "schedulingAreaIds": {
            "default": [],
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          }
        },
        "required": [
          "name",
          "email",
          "accessProfile",
          "permissions"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "accessProfile": {
            "type": "string",
            "enum": [
              "admin",
              "professional"
            ]
          },
          "servesClients": {
            "default": false,
            "type": "boolean"
          },
          "permissions": {
            "maxItems": 24,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "admin.users.read",
                "admin.users.create",
                "admin.users.update",
                "admin.users.delete",
                "admin.users.trash.read",
                "admin.users.trash.restore",
                "admin.users.trash.purge",
                "admin.users.access_link.resend",
                "admin.users.audit.read",
                "admin.permission_templates.read",
                "admin.permission_templates.create",
                "admin.permission_templates.update",
                "admin.permission_templates.delete",
                "admin.permission_templates.audit.read",
                "admin.tags.read",
                "admin.tags.create",
                "admin.tags.update",
                "admin.tags.delete",
                "admin.tags.trash.read",
                "admin.tags.trash.restore",
                "admin.tags.trash.purge",
                "admin.tags.audit.read",
                "admin.audit.read",
                "admin.usage.read"
              ]
            }
          },
          "areaIds": {
            "default": [],
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          },
          "serviceIds": {
            "default": [],
            "maxItems": 2000,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          },
          "schedulingAreaIds": {
            "default": [],
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          }
        },
        "required": [
          "name",
          "accessProfile",
          "permissions"
        ]
      },
...
    }
  },
  "security": [
    {
      "cookie": []
    }
  ]
}