# packages/api-client/generated/models/UpdateUserDto.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const updateUserDtoAccessProfileEnum = {
    "admin": "admin",
    "professional": "professional"
} as const;

export type UpdateUserDtoAccessProfileEnum = (typeof updateUserDtoAccessProfileEnum)[keyof typeof updateUserDtoAccessProfileEnum];

export const updateUserDtoPermissionsEnum = {
    "admin.users.read": "admin.users.read",
    "admin.users.create": "admin.users.create",
    "admin.users.update": "admin.users.update",
    "admin.users.delete": "admin.users.delete",
    "admin.users.trash.read": "admin.users.trash.read",
    "admin.users.trash.restore": "admin.users.trash.restore",
    "admin.users.trash.purge": "admin.users.trash.purge",
    "admin.users.access_link.resend": "admin.users.access_link.resend",
    "admin.users.audit.read": "admin.users.audit.read",
    "admin.permission_templates.read": "admin.permission_templates.read",
    "admin.permission_templates.create": "admin.permission_templates.create",
    "admin.permission_templates.update": "admin.permission_templates.update",
    "admin.permission_templates.delete": "admin.permission_templates.delete",
    "admin.permission_templates.audit.read": "admin.permission_templates.audit.read",
    "admin.tags.read": "admin.tags.read",
    "admin.tags.create": "admin.tags.create",
    "admin.tags.update": "admin.tags.update",
    "admin.tags.delete": "admin.tags.delete",
    "admin.tags.trash.read": "admin.tags.trash.read",
    "admin.tags.trash.restore": "admin.tags.trash.restore",
    "admin.tags.trash.purge": "admin.tags.trash.purge",
    "admin.tags.audit.read": "admin.tags.audit.read",
    "admin.audit.read": "admin.audit.read",
    "admin.usage.read": "admin.usage.read"
} as const;

export type UpdateUserDtoPermissionsEnum = (typeof updateUserDtoPermissionsEnum)[keyof typeof updateUserDtoPermissionsEnum];

export type UpdateUserDto = {
    /**
     * @minLength 1
     * @type string
    */
    name: string;
    /**
     * @type string
    */
    accessProfile: UpdateUserDtoAccessProfileEnum;
    /**
     * @default false
     * @type boolean | undefined
    */
    servesClients?: boolean;
    /**
     * @type array
    */
    permissions: UpdateUserDtoPermissionsEnum[];
    /**
     * @type array | undefined
    */
    areaIds?: string[];
    /**
     * @type array | undefined
    */
    serviceIds?: string[];
    /**
     * @type array | undefined
    */
    schedulingAreaIds?: string[];
};